requirements.txt
asgiref==3.4.1
boto3==1.18.20
botocore==1.21.20
bottle==0.12.19
coverage==5.5
CProfileV==1.0.7
Django==3.2.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-storages==1.11.1
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==13.4.0
jmespath==0.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
s3transfer==0.5.0
serpy==0.3.1
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.2.0
urllib3==1.26.6